split classes into dif files
