+++ index.html
@@ -9,7 +9,7 @@
 <body>
     <script src="lib/phaser.js"></script>
     <script src="src/scenes/main_game_scene.js"></script>
-    <script src="main.js"></script>
+    <script type="module" src="main.js"></script>
    
 </body>
 </html>
